Crate ddc_hi

High level DDC/CI monitor controls.

Example

use ddc_hi::{Ddc, Display};

for mut display in Display::enumerate() {
    display.update_capabilities().unwrap();
    println!("{:?} {}: {:?} {:?}",
        display.info.backend, display.info.id,
        display.info.manufacturer_id, display.info.model_name
    );
    if let Some(feature) = display.info.mccs_database.get(0xdf) {
        let value = display.handle.get_vcp_feature(feature.code).unwrap();
        println!("{}: {:?}", feature.name.as_ref().unwrap(), value);
    }
}

Structs

Display
An active handle to a connected display.

DisplayInfo
Identifying information about an attached display.

TimingMessage

VcpValue
VCP Value

Enums

Backend
Identifies the backend driver used to communicate with a display.

Handle
A handle allowing communication with a display

Query
A query to filter out matching displays.

VcpValueType
VCP feature type.

Traits

Ddc
A high level interface to DDC commands.

DdcHost
A DDC host is able to communicate with a DDC device such as a display.

DdcTable
Table commands can read and write arbitrary binary data to a VCP feature.

Type Definitions

FeatureCode
VCP feature code