Enum ddc_hi::VcpValueType

#[repr(u8)]

pub enum VcpValueType {
    SetParameter,
    Momentary,
}

VCP feature type.

Variants
SetParameter

Sending a command of this type changes some aspect of the monitor's operation.

Momentary

Sending a command of this type causes the monitor to initiate a self-timed operation and then revert to its original state.

Examples include display tests and degaussing.