Trait ddc_hi::DdcHost

pub trait DdcHost {
    type Error;
    fn sleep(&mut self) { ... }
}

A DDC host is able to communicate with a DDC device such as a display.

Associated Types

type Error

An error that can occur when communicating with a DDC device.

Usually impls From<ErrorCode>.

Provided Methods

fn sleep(&mut self)

Wait for any previous commands to complete.

The DDC specification defines delay intervals that must occur between execution of two subsequent commands, this waits for the amount of time remaining since the last command was executed. This is normally done internally and shouldn't need to be called manually unless synchronizing with an external process or another handle to the same device. It may however be desireable to run this before program exit.